크롤러 코드 이전 및 오류 수정 (#114)
