+++ backend/problem/utils.py
@@ -1,23 +1,15 @@
 import json
 import random
 import re
-import time
 
 import dramatiq
 from functools import lru_cache
 
 from django.db import transaction
 from django.db.models import Min, F
-from selenium.common.exceptions import NoSuchElementException
-from selenium.webdriver.common.by import By
 
-from announcement.models import LinkAnnouncement
 from .models import Problem, get_default_week_info
-from utils.constants import Difficulty, LINK_NOTICE_LIMIT, LINK_NOTICE_SCRAPING_URL
-from selenium import webdriver
-import logging
-
-logger = logging.getLogger(__name__)
+from utils.constants import Difficulty
 
 TEMPLATE_BASE = """//PREPEND BEGIN
 {}
@@ -124,115 +116,3 @@ def update_bonus_problem():
             problem.is_bonus = True
             problem.save(update_fields=['is_bonus'])
 
-
-def is_element_exist(item, by, locator):
-    try:
-        item.find_element(by, locator)
-        return True
-    except NoSuchElementException:
-        return False
-
-
-def diff_link_announcement(first_id):
-    base_objs = LinkAnnouncement.objects.all()[:LINK_NOTICE_LIMIT].first().la_id
-    return int(first_id) - base_objs
-
-
-def update_link_announcement(rows, diff_cnt):
-    with transaction.atomic():
-        for k, v in enumerate(rows):
-            link_announcement = LinkAnnouncement(
-                la_id=v['id'],
-                title=v['title'],
-                url=v['url'],
-                create_time=v['create_time'],
-                new_flag=v['new_flag']
-            )
-            if k <= diff_cnt - 1:
-                link_announcement.save()
-            else:
-                link_announcement.save(update_fields=['new_flag'])
-
-def scrap_and_update():
-    scrap_link_announcement.send()
-@dramatiq.actor()
-def scrap_link_announcement():
-    # 브라우저를 열지 않고 실행
-    options = webdriver.ChromeOptions()
-    options.add_argument("headless")
-
-    # 공지사항 접속
-    browser = webdriver.Chrome(options=options)
-    browser.get(LINK_NOTICE_SCRAPING_URL)
-    time.sleep(7)
-
-    # tbody(테이블 바디)가 있는지 확인
-    if not is_element_exist(browser, By.CSS_SELECTOR, "tbody"):
-        browser.close()
-        return
-
-    # 첫번째 tr이 headline인지 아닌지에 따른 분기처리
-    first_element = browser.find_element(By.CSS_SELECTOR, "tbody > tr:nth-child(1)")
-    if "headline" in first_element.get_attribute("class"):
-        first_id = browser.find_element(By.CSS_SELECTOR, "tbody > tr:nth-child(2)").find_element(By.CLASS_NAME,
-                                                                                                 "_artclTdNum").text
-    else:
-        first_id = first_element.find_element(By.CLASS_NAME, "_artclTdNum").text
-
-    diff_cnt = LINK_NOTICE_LIMIT if LinkAnnouncement.objects.count() == 0 else diff_link_announcement(first_id)
-    if diff_cnt == 0:
-        logger.info("There is no additional notice")
-
-    time.sleep(1)
-    rows = browser.find_elements(By.CSS_SELECTOR, "tbody > tr")
-
-    # 첫번째 tr이 headline인지 아닌지에 따른 분기처리
-    if "headline" in rows[0].get_attribute("class"):
-        rows = rows[1:LINK_NOTICE_LIMIT + 1]
-    else:
-        rows = rows[:LINK_NOTICE_LIMIT]
-
-    target_list = []
-
-    cnt = 0
-    for row in rows:
-        # 번호
-        id = row.find_element(By.CLASS_NAME, "_artclTdNum").text
-        # 작성일
-        create_time = row.find_element(By.CLASS_NAME, "_artclTdRdate").text
-        # 새 글
-        new_flag = is_element_exist(row, By.CLASS_NAME, "newArtcl")
-        logger.info("{} {} {}".format(id, create_time, new_flag))
-        title = curr_url = None
-
-        if cnt < diff_cnt:
-            # 내부 컨텐츠 진입
-            a_title_element = row.find_element(By.CLASS_NAME, "artclLinkView")
-            a_title_element.click()
-            time.sleep(3)
-
-            # 전체 제목
-            title = browser.find_element(By.CLASS_NAME, "artclViewTitle").text
-            # 링크
-            curr_url = browser.current_url
-
-            browser.back()
-            browser.implicitly_wait(1)
-
-        cnt += 1
-        target_list.append({
-            "id": int(id),
-            "title": title,
-            "url": curr_url,
-            "create_time": create_time,
-            "new_flag": new_flag
-        })
-
-    logger.info(target_list)
-
-    if len(target_list) > 0:
-        update_link_announcement(target_list, diff_cnt)
-
-    logger.info("Scrap and update done")
-
-    browser.close()
